Split out stencil shadow drawing into separate file
git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@7226 49c82d7f-2e2a-0410-a16f-ae8f201b507f
