+++ darkmodNT.vcxproj
@@ -1178,6 +1178,7 @@
     <ClCompile Include="renderer\draw_arb2.cpp" />
     <ClCompile Include="renderer\draw_common.cpp" />
     <ClCompile Include="renderer\draw_glsl.cpp" />
+    <ClCompile Include="renderer\draw_stencil.cpp" />
     <ClCompile Include="renderer\FrameBufer.cpp" />
     <ClCompile Include="renderer\GuiModel.cpp" />
     <ClCompile Include="renderer\Image_files.cpp" />
